Final gloss on parser refactor
This commit represents the final gloss on the parser refactor.  It cleans
up some documentation and corrects a missed issue where UnaryFactory and
BinaryFactory did not conform to the factory functions--they were missing
the IParser and IPushBackLexer arguments.
