+++ parser/visualize/doc.go
@@ -12,11 +12,11 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
-// Package parser is a subpackage of ptk that contains an
-// implementation of the IParser, an interface describing parsers,
-// along with related support types and code such as Node.  A parser
-// is an object that can be queried to extract expressions and
-// statements from a lexer; generally, a parser takes a lexer and
-// builds up an abstract syntax tree from the tokens generated by the
-// lexer.
-package parser
+// Package visualize contains a Visualize function that may be used to
+// generate a printable tree representation of a nested set of Node
+// instances.  This may be used to visualize an abstract syntax tree,
+// as generated by the parser, but may also be used to visualize other
+// tree-like structures, as long as the elements of those structures
+// implement the parser.Node interface.  Different profiles exist that
+// may be used to alter the visual appearance of the visualized tree.
+package visualize
